requirements.txt
streamlit
pulp
scikit-learn